# packages/server/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/stream": {
            "delete": {
                "description": "Stop playback of the current stream (if any). Note that a client may continue playing until its local buffer is empty.",
                "operationId": "StreamController_stop",
                "parameters": [],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "A query or route parameter, the payload or a header was malformed and did not pass validation."
                    },
                    "500": {
                        "description": "An unexpected error occurred."
                    }
                },
                "security": [
                    {}
                ],
                "summary": "Stop playback.",
                "tags": [
                    "stream"
                ]
            },
            "get": {
                "description": "Get the track currently being streamed (if any).",
                "operationId": "StreamController_nowPlaying",
                "parameters": [],
                "responses": {
                    "400": {
                        "description": "A query or route parameter, the payload or a header was malformed and did not pass validation."
                    },
                    "404": {
                        "description": "No channel is currently being streamed."
                    },
                    "500": {
                        "description": "An unexpected error occurred."
                    },
                    "default": {
                        "description": "The operation completed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NowPlayingDTO_Output"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {}
                ],
                "summary": "Get track being streamed.",
                "tags": [
                    "stream"
                ]
            }
        },
        "/stream/{networkKey}/{channelKey}": {
            "get": {
                "description": "Start playback of the channel of the given network with the given key, if it exists. The stream is the raw audio without any IceCast metadata. Any previously started stream is terminated.",
                "operationId": "StreamController_play",
                "parameters": [
                    {
                        "name": "channelKey",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "example": "trance",
                            "type": "string"
                        }
                    },
                    {
                        "name": "networkKey",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "example": "di",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The operation completed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "A query or route parameter, the payload or a header was malformed and did not pass validation."
                    },
                    "404": {
                        "description": "The channel with the given key was not found."
                    },
                    "500": {
                        "description": "An unexpected error occurred."
                    }
                },
                "security": [
                    {}
                ],
                "summary": "Start streaming a channel.",
                "tags": [
                    "stream"
                ]
            }
        }
    },
    "info": {
        "title": "My API",
        "description": "This is my API.",
        "version": "1.0",
        "contact": {
            "name": "Raphael Pigulla",
            "url": "https://de.linkedin.com/in/pigulla",
            "email": "raphael@pigulla.net"
        },
        "license": {
            "name": "ISC",
            "url": "https://spdx.org/licenses/ISC"
        }
    },
...
    "servers": [
        {
            "url": "http://localhost.com:3000"
        }
    ],
    "components": {
        "schemas": {
...
            "NowPlayingDTO_Output": {
                "type": "object",
                "properties": {
                    "track": {
                        "type": "string"
                    },
                    "network": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 9007199254740991
                            },
                            "key": {
                                "type": "string",
                                "pattern": "^[a-z_0-9]+$"
                            },
                            "name": {
                                "type": "string",
                                "minLength": 1
                            },
                            "url": {
                                "type": "string",
                                "format": "uri"
                            }
                        },
                        "required": [
                            "id",
                            "key",
                            "name",
                            "url"
                        ],
                        "additionalProperties": false
                    },
                    "channel": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 9007199254740991
                            },
                            "key": {
                                "type": "string",
                                "pattern": "^[a-z_0-9]+$"
                            },
                            "networkId": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 9007199254740991
                            },
                            "name": {
                                "type": "string",
                                "minLength": 1
                            },
                            "director": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "key",
                            "networkId",
                            "name",
                            "director",
                            "description"
                        ],
                        "additionalProperties": false
                    }
                },
                "required": [
                    "track",
                    "network",
                    "channel"
                ]
            }
        }
    }
}